+++ spatial_transformation/__init__.py
@@ -1,4 +1,4 @@
 # SPDX-License-Identifier: BSD-3-Clause
 """Utility package for working with spatial transformations."""
 
-from .transform import Transformation, Rotation
+from ._transformation import Transformation, Rotation
